+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [CalVer](https://calver.org/).
 
-## [UNRELEASED] - 2022-17-11
+## [2022.11.2] - 2022-17-11
 ### Changed
 - Changed from noUiSlider to native HTML range input in these modules. noUiSlider packages is no longer installed:
   - Base layers
   - Advanced info
   - Conflict
+- jQuery upgraded to 3.6.1
 
 ## [2022.11.1] - 2022-16-11
 ### Changed